Move WebSocket tests from test_coordinator to test_button per review
